MuPDF 1.12 (#577)
* Updated to MuPDF 1.12

* mupdf_new_pixmap_with_data missed parameters

* Make all parameters of fz_get_pixmap_from_image available, remove the unused with and height

* Drop buffer and image as soon as possible as per poire-z's recommendation

* encrypted zip patch for MuPDF 1.12

* external fonts patch for MuPDF 1.12

* Use libjpeg as a shared library

* Wrong patch number

* git.ghostscript.com is slow, so we use the official mirror on GitHub

* html-font patch, just enough to test html & epub rendering

* Regenerated mupdf_h.lua with ffi-cdecl

* fz_new_text_page has been renamed to fz_new_stext_page

* Wrong mupdf_new_text_page parameters, mupdf_new_text_sheet is no longer needed

* getPageText works again

* Adding higlights almost works (width is 50% off), PDF saving works

Requires changes to PdfDocument:saveHighlight too (that is outside of
base).

* Do not select bullet points

* The incompatible bb error messages show the blitbuffer types

* Support blitting from RGB24 to RGB32 in BB_alpha_blit_from

* getPageText returned everything twice

* Unit test update

* 'SIZE_MAX' was undeclared on Android

+++ wrap-mupdf.c
@@ -17,8 +17,10 @@
 */
 
 #include <stdio.h>
+#include <stdlib.h>
 #include <math.h>
 #include <stddef.h>
+#include <string.h>
 #include "wrap-mupdf.h"
 
 static double LOG_TRESHOLD_PERC = 0.05; // 5%
@@ -65,7 +67,7 @@ static void log_size(char *funcName){
 }
 
 static void *
-my_malloc_default(void *opaque, unsigned int size)
+my_malloc_default(void *opaque, size_t size)
 {
     struct header * h = malloc(size + sizeof(header));
     if (h == NULL)
@@ -97,7 +99,7 @@ my_free_default(void *opaque, void *ptr)
 }
 
 static void *
-my_realloc_default(void *opaque, void *old, unsigned int size)
+my_realloc_default(void *opaque, void *old, size_t size)
 {
     void * newp;
     if (old==NULL) { //practically, it's a malloc
